Move thumbnail loading into download module
