+++ mobile/apps/photos/lib/services/wrapped/wrapped_media_preloader.dart
@@ -3,8 +3,8 @@ import "dart:async";
 import "package:logging/logging.dart";
 import "package:photos/db/files_db.dart";
 import "package:photos/models/file/file.dart";
+import "package:photos/module/download/thumbnail.dart";
 import "package:photos/services/wrapped/models.dart";
-import "package:photos/utils/thumbnail_util.dart";
 
 /// Preloads the media referenced by Wrapped cards so the viewer can render
 /// without waiting on database and thumbnail fetches.
