+++ mobile/apps/photos/lib/ui/home/memories/full_screen_memory.dart
@@ -22,6 +22,7 @@ import "package:photos/models/file/extensions/file_props.dart";
 import "package:photos/models/file/file.dart";
 import "package:photos/models/file/file_type.dart";
 import "package:photos/models/memories/memory.dart";
+import "package:photos/module/download/thumbnail.dart";
 import "package:photos/service_locator.dart";
 import "package:photos/services/memory_share_service.dart";
 import "package:photos/services/smart_memories_service.dart";
@@ -40,7 +41,6 @@ import "package:photos/ui/viewer/gallery/jump_to_date_gallery.dart";
 import "package:photos/utils/dialog_util.dart";
 import "package:photos/utils/file_util.dart";
 import "package:photos/utils/share_util.dart";
-import "package:photos/utils/thumbnail_util.dart";
 
 //There are two states of variables that FullScreenMemory depends on:
 //1. The list of memories
