+++ mobile/apps/photos/lib/ui/viewer/search/result/edit_contact_page.dart
@@ -15,6 +15,7 @@ import "package:photos/models/file/file.dart";
 import "package:photos/models/ml/face/person.dart";
 import "package:photos/models/search/generic_search_result.dart";
 import "package:photos/models/search/search_constants.dart";
+import "package:photos/module/download/thumbnail.dart";
 import "package:photos/services/machine_learning/face_ml/face_filtering/face_filtering_constants.dart";
 import "package:photos/services/machine_learning/face_ml/person/person_service.dart";
 import "package:photos/services/machine_learning/ml_result.dart";
@@ -34,7 +35,6 @@ import "package:photos/ui/viewer/search/result/contact_photo_adjust_page.dart";
 import "package:photos/ui/viewer/search/result/contact_photo_picker_sheet.dart";
 import "package:photos/utils/dialog_util.dart";
 import "package:photos/utils/face/face_thumbnail_cache.dart";
-import "package:photos/utils/thumbnail_util.dart";
 
 class EditContactPage extends StatefulWidget {
   final int contactUserId;
